requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
